shifted logic to plugins

+++ mirror_plugin/src/lib.rs
@@ -16,7 +16,6 @@ pub unsafe extern "C" fn process_image(
     rgba_data: *mut c_uchar,
     params: *const c_char,
 ) -> i32 {
-
     let Some(data_size) = (width as usize)
         .checked_mul(height as usize)
         .and_then(|res| res.checked_mul(PIXEL_BYTES as usize))
@@ -26,14 +25,18 @@ pub unsafe extern "C" fn process_image(
 
     // SAFETY: rgba_data must have at least data_size bytes
     let rgba_data_slice = unsafe { std::slice::from_raw_parts_mut(rgba_data, data_size) };
-    mirror_horizontal(rgba_data_slice, width, height);
+    mirror_horizontal(rgba_data_slice, width as usize, height as usize);
     mirror_vertical(rgba_data_slice, width, height);
     0
 }
 
 fn mirror_vertical(buf: &mut [u8], width: u32, height: u32) {
     let stride = width * PIXEL_BYTES;
-    assert_eq!(buf.len() as u32, stride * height, "invalid RGBA buffer size");
+    assert_eq!(
+        buf.len() as u32,
+        stride * height,
+        "invalid RGBA buffer size"
+    );
 
     for y in 0..height / 2 {
         let top_start = y * stride;
@@ -45,22 +48,21 @@ fn mirror_vertical(buf: &mut [u8], width: u32, height: u32) {
     }
 }
 
-fn mirror_horizontal(buf: &mut [u8], width: u32, height: u32) {
+fn mirror_horizontal(buf: &mut [u8], width: usize, height: usize) {
     let stride = width * 4;
-    assert_eq!(buf.len() as u32, stride * height, "invalid RGBA buffer size");
+    assert_eq!(buf.len(), stride * height, "invalid RGBA buffer size");
 
-    for line in buf.chunks_exact_mut(stride as usize) {
+    for line in buf.chunks_exact_mut(stride) {
         for x in 0..width / 2 {
             let a = x * 4;
             let b = (width - 1 - x) * 4;
-            line.swap(a as usize, b as usize);
-            line.swap((a + 1) as usize, (b + 1) as usize);
-            line.swap((a + 2) as usize, (b + 2) as usize);
-            line.swap((a + 3) as usize, (b + 3) as usize);
+            line.swap(a, b);
+            line.swap((a + 1), (b + 1));
+            line.swap((a + 2), (b + 2));
+            line.swap((a + 3), (b + 3));
         }
     }
 }
-
 
 #[cfg(test)]
 mod tests {
